# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Documentación automática del backend de Taskly",
        "title": "Taskly API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:7000",
    "basePath": "/",
    "paths": {
...
        "/equipos/{id}/tareas": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna todas las tareas asociadas a un equipo específico, incluyendo el nombre del usuario asignado (si existe).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tareas"
                ],
                "summary": "Listar tareas de un equipo",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "ID del equipo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tareas obtenida exitosamente",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TareaConAsignadoResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "ID de equipo inválido",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Equipo no encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea una tarea dentro de un equipo específico y la inicializa en estado \"backlog\". Opcionalmente se puede asignar a un miembro.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tareas"
                ],
                "summary": "Crear una nueva tarea en un equipo",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "ID del equipo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de la tarea a crear",
                        "name": "datos",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Tarea"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tarea creada exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.Tarea"
                        }
                    },
                    "400": {
                        "description": "ID de equipo inválido o datos incorrectos",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Equipo no encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/tareas/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna los detalles completos de una tarea específica identificada por su ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tareas"
                ],
                "summary": "Obtener una tarea por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 42,
                        "description": "ID de la tarea",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea obtenida exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.Tarea"
                        }
                    },
                    "400": {
                        "description": "ID de tarea inválido",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permite modificar los campos de una tarea (título, descripción, asignado, etc.). Solo campos enviados serán actualizados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tareas"
                ],
                "summary": "Actualizar una tarea existente",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 42,
                        "description": "ID de la tarea",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a actualizar",
                        "name": "datos",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Tarea"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea actualizada exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.Tarea"
                        }
                    },
                    "400": {
                        "description": "ID inválido o datos incorrectos",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Elimina permanentemente una tarea del sistema. Esta acción no se puede deshacer.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tareas"
                ],
                "summary": "Eliminar una tarea",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 42,
                        "description": "ID de la tarea a eliminar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "ID de tarea inválido",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Permisos insuficientes para eliminar",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tareas/{id}/estado": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza únicamente el campo `estado` de una tarea. Valores permitidos: \"backlog\", \"pendiente\", \"en_progreso\", \"completada\".",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tareas"
                ],
                "summary": "Cambiar el estado de una tarea",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 42,
                        "description": "ID de la tarea",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevo estado de la tarea",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateEstadoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado actualizado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "ID inválido, datos incorrectos o estado no permitido",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Datos inválidos"
                },
                "message": {
                    "type": "string",
                    "example": "Campo requerido"
                }
            }
        },
...
        "models.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operación exitosa"
                }
            }
        },
...
        "models.Tarea": {
            "type": "object",
            "properties": {
                "asignadoid": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "descripcion": {
                    "type": "string"
                },
                "equipoid": {
                    "type": "integer"
                },
                "estado": {
                    "type": "string"
                },
                "fechaLimite": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "prioridad": {
                    "type": "string"
                },
                "titulo": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.TareaConAsignadoResponse": {
            "description": "Respuesta de tarea que incluye el nombre del miembro asignado (si existe)",
            "type": "object",
            "properties": {
                "asignado_id": {
                    "type": "integer",
                    "example": 5
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-05-28T10:00:00Z"
                },
                "descripcion": {
                    "type": "string",
                    "example": "Corregir error en autenticación"
                },
                "equipo_id": {
                    "type": "integer",
                    "example": 3
                },
                "estado": {
                    "type": "string",
                    "example": "en_progreso"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nombreAsignado": {
                    "type": "string",
                    "example": "Ana López"
                },
                "titulo": {
                    "type": "string",
                    "example": "Fix login bug"
                }
            }
        },
        "models.UpdateEstadoRequest": {
            "description": "Solo contiene el campo estado con valores permitidos",
            "type": "object",
            "properties": {
                "estado": {
                    "type": "string",
                    "enum": [
                        "backlog",
                        "pendiente",
                        "en_progreso",
                        "completada"
                    ],
                    "example": "en_progreso"
                }
            }
        },
...
    }
}